+++ tests/Component/CollectionTest.php
@@ -88,12 +88,36 @@ class CollectionTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * Tests if getComponent() returns null.
+     * Tests if get() returns null.
      *
      * @return void
      */
     public function testGetNullComponent()
     {
         $this->assertNull($this->components->getContainer());
+    }
+
+    /**
+     * Tests if get('test') returns an UnexpectedValueException.
+     *
+     * @return void
+     */
+    public function testGetUnknownComponentType()
+    {
+        $this->setExpectedException('UnexpectedValueException');
+
+        $this->components->get('test');
+    }
+
+    /**
+     * Tests if set('test', null) returns an UnexpectedValueException.
+     *
+     * @return void
+     */
+    public function testSetUnknownComponentType()
+    {
+        $this->setExpectedException('UnexpectedValueException');
+
+        $this->components->set('test', null);
     }
 }